:minibus: Rename files properly
